Moved progress-related utilities, and improved "locked" UI
Now, the titles of locked missions are grayed out, a lock icon is added, and no gems/rooms are highlighted when all missions on a level are locked.
