+++ src/code/fv-components/navigation-dialog.js
@@ -3,6 +3,7 @@ import t from '../utilities/translate';
 import GemSetView from './gem-set';
 import VenturePadView from './venture-pad';
 import AuthoringUtils from '../utilities/authoring-utils';
+import ProgressUtils from '../utilities/progress-utils';
 
 export default class NavigationDialogView extends React.Component {
 
@@ -19,7 +20,7 @@ export default class NavigationDialogView extends React.Component {
     let level = props.routeSpec && props.routeSpec.level;
 
     if (typeof level !== "number") {
-      level = AuthoringUtils.getCurrentChallengeFromGems(props.authoring, props.gems).level;
+      level = ProgressUtils.getCurrentChallengeFromGems(props.authoring, props.gems).level;
     }
 
     this.state = {
@@ -29,18 +30,17 @@ export default class NavigationDialogView extends React.Component {
 
   render() {
     let {gems, onNavigateToChallenge, onHideMap, authoring} = this.props,
-        currChallengeRoute = AuthoringUtils.getCurrentChallengeFromGems(authoring, gems),
-        currChallengeRouteForLevel = AuthoringUtils.getCurrentChallengeFromGems(authoring, gems, this.state.level);
+        currChallengeRouteForLevel = ProgressUtils.getCurrentChallengeFromGems(authoring, gems, this.state.level);
 
     let gemSets = [];
 
     for (let mission = 0; mission < AuthoringUtils.getMissionCount(authoring, this.state.level); mission++) {
       let isCurrMission = mission === currChallengeRouteForLevel.mission,
-          isLocked = this.state.level > currChallengeRoute.level 
-                  || (this.state.level === currChallengeRoute.level && mission > currChallengeRouteForLevel.mission );
-      gemSets.push(<div id={"gem-label-" + mission} className="gem-set-label" key={"label-" + mission}>
+          isLocked = ProgressUtils.isMissionLocked(gems, authoring, this.state.level, mission);
+      gemSets.push(<div id={"gem-label-" + mission} className="gem-set-label" key={"label-" + mission} style={isLocked ? {opacity: .5} : null}>
                      {"Mission " + (this.state.level + 1) + "." + (mission + 1) + ":"}
                    </div>);
+      gemSets.push(<div className="mission-lock" style={isLocked ? null : {visibility: "hidden"}}></div>);
       gemSets.push(<GemSetView level={this.state.level} mission={mission} challenge={isCurrMission ? currChallengeRouteForLevel.challenge : null}
                                challengeCount={AuthoringUtils.getChallengeCount(authoring, this.state.level, mission)}
                                gems={gems} key={"gem-set-" + mission} isLocked={isLocked}
@@ -74,8 +74,9 @@ export default class NavigationDialogView extends React.Component {
       </div>
     );
 
+    let highlightedRoomLocked = ProgressUtils.isMissionLocked(gems, authoring, currChallengeRouteForLevel.level, currChallengeRouteForLevel.mission);
     return <VenturePadView title={t("~VENTURE.MAP")} screen={screen} onClickOutside={onHideMap} onNavigateToChallenge={onNavigateToChallenge}
-                           authoring={authoring} roomHighlightRoute={currChallengeRouteForLevel}></VenturePadView>;
+                           authoring={authoring} roomHighlightRoute={highlightedRoomLocked ? null : currChallengeRouteForLevel}></VenturePadView>;
   }
 }
