+++ src/code/reducers/helpers/load-state-from-authoring.js
@@ -2,6 +2,7 @@ import templates from '../../templates';
 import GeneticsUtils from '../../utilities/genetics-utils';
 import { range, shuffle, assign } from 'lodash';
 import AuthoringUtils from '../../utilities/authoring-utils';
+import ProgressUtils from '../../utilities/progress-utils';
 import { notificationType } from '../../modules/notifications';
 
 /**
@@ -200,8 +201,8 @@ export function loadHome(state, authoring, showMissionEndDialog) {
   let room = "home",
       roomInfo = (authoring && authoring.rooms) ? authoring.rooms[room] : {},
       location = {id: room, ...roomInfo},
-      { level, mission } = AuthoringUtils.getCurrentChallengeFromGems(authoring, state.gems),
-      missionStarted = AuthoringUtils.isMissionStarted(state.gems, level, mission),
+      { level, mission } = ProgressUtils.getCurrentChallengeFromGems(authoring, state.gems),
+      missionStarted = ProgressUtils.isMissionStarted(state.gems, level, mission),
       dialogDefs = authoring.application.levels[level].missions[mission].dialog,
       messages = [];
