+++ src/stylus/fablevision/navigation-dialog.styl
@@ -31,3 +31,9 @@
       background-image: url('../resources/fablevision/venture-pad/right_nav_button.png')
   #gem-label-0
     margin-top: 20px
+  .mission-lock
+    width: 56px
+    height: 52px
+    background-image: url('../resources/fablevision/venture-pad/lock.png')
+    margin-top: -60px
+    margin-left: 290px
